requirements.txt
requests>=2.28.0
gtfs-realtime-bindings>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
pyarrow>=14.0.0
